Testing Update
The first test in GitHub actions seems to fail, despite passing in the local mocha chai testing environment.
This test has been duplicated to determine if the issue is a connection or testing issue.
Additionally the testing process times out, an exit process function has been added to the code.